pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "saasmint-core"
version = "0.4.0"
description = "Django backend for SaasMint"
requires-python = ">=3.12"
dependencies = [
    "saasmint-core-lib",
    "django>=6.0.3",
    "djangorestframework>=3.15",
    "django-cors-headers>=4.7",
    "psycopg[binary]>=3.3.3",
    "celery[redis]>=5.6.3",
    # redis is pinned by celery's kombu dep to <6.5; redis 7.x requires a kombu/celery upgrade first
    "pydantic-settings>=2.13.1",
    "django-hijack>=3.7.6",     # user impersonation in admin
    "whitenoise>=6.12.0",
    "pyjwt[crypto]>=2.12.1",
    "uvicorn[standard]>=0.42.0",
    "pydantic[email]>=2.12.5",
    "supabase>=2.28.3",
    "drf-spectacular>=0.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "testcontainers[postgres]>=4.0",
    "mypy>=1.14",
    "django-stubs>=6.0",
    "djangorestframework-stubs>=3.15",
    "ruff>=0.14",
    "pip-audit>=2.7",
]

[tool.hatch.build.targets.wheel]
packages = ["config", "apps", "middleware"]

[tool.uv.sources]
saasmint-core-lib = { path = "core", editable = true }

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
testpaths = ["core/tests", "apps/users/tests", "apps/billing/tests", "apps/orgs/tests", "apps/dashboard/tests", "apps/admin_panel/tests", "config/tests", "scripts/tests"]
filterwarnings = [
    "ignore:No directory at.*staticfiles:UserWarning",
    "ignore:Error when trying to teardown test databases:pytest.PytestWarning",
]
addopts = "--cov-config=pyproject.toml -p no:cacheprovider"

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
strict = true
exclude = [".venv", "node_modules", "__pycache__", "tests", "migrations", "apps/users/management", "apps/admin_panel", "apps/dashboard"]

[[tool.mypy.overrides]]
module = ["celery", "celery.*"]
ignore_missing_imports = true
follow_untyped_imports = false

[tool.django-stubs]
django_settings_module = "config.settings.test"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ANN", "S", "B", "A", "C4", "RUF", "DJ"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "ANN"]
"apps/*/tests/**" = ["S101", "ANN", "S105"]
"scripts/tests/**" = ["S101", "ANN"]
"config/tests/**" = ["S101", "ANN"]
"*/migrations/**" = ["ANN", "RUF", "E501"]

[tool.coverage.run]
omit = [
    "config/settings/dev.py",
    "config/settings/prod.py",
    "config/asgi.py",
    "config/wsgi.py",
]

[dependency-groups]
dev = [
    "pytest-anyio>=0.0.0",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.12.0",
]
